pd_vse/requirements.txt
fastapi
fastapi-sqlalchemy
SQLAlchemy
uvicorn[standard]
python-decouple
behave
yml
opencv-python
cython
wget
schedule
kafka-python
aiohttp
python-multipart
celery
flower
redis
pika
timeloop